.history/requirements_20240315144350.txt
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
pandas
urllib3==2.0.4
requests==2.31.0
flask_mail
Flask-Mail
json5==0.9.14
openpyxl==3.1.2
pybase64==1.3.1
redis-py